modularize compare/search logic
